refactor(calculator): extract model resolution into resolve.py
